% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custbarplot.R
\name{custbarplot}
\alias{custbarplot}
\title{Make a bar plot of responses on Likert scale}
\usage{
custbarplot(data, xvar, grpvar, withingrp = FALSE, scalepoints = 7,
  decimals = TRUE, scale = "Agree", ...)
}
\arguments{
\item{data}{Name of dataset (must be in wide format, i.e., one row per participant)}

\item{xvar}{Name of variable of interest measured on Likert scale}

\item{grpvar}{Name of variable of type factor by which bars should be split}

\item{withingrp}{Logical value indicating whether or not percentages should be calculated within groups}

\item{scalepoints}{Numeric value indicating number of scalepoints in Likert scale (at present, can take on value of 5 or 7)}

\item{decimals}{Logical value indicating whether or not xvar is on a non-integer scale}

\item{scale}{Character value indicating scale labels (at present, can take on values of "Agree", "Quant", or "True")}

\item{...}{Other arguments to pass on to ggplot}
}
\value{
A barplot showing the counts and percentages of all values of xvar
}
\description{
Make a bar plot of responses on Likert scale
}
\examples{
mtcars$am <- as.factor(mtcars$am)
custbarplot(mtcars, xvar = gear, grpvar = am, withingrp = TRUE, scalepoints = 5, decimals = FALSE)
}
