% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_student_data.R
\docType{data}
\name{random_student_data}
\alias{random_student_data}
\title{Dataset containing info on students' demographic characteristics and math-related attitudes and perceptions}
\format{A dataframe with 500 rows and 8 variables:
\describe{
  \item{studentID}{ID number of participant}
  \item{sex}{Sex of participant}
  \item{race}{Race of participant}
  \item{age}{Age of participant}
  \item{grade_level}{Grade level of participant}
  \item{math_motivation}{Degree to which participant felt motivated to learn math at end of two-week period}
  \item{perceived_math_meaning_week1}{Degree to which participant agreed that teacher helped them find a meaningful reason to learn math during week 1}
  \item{perceived_math_meaning_week2}{Degree to which participant agreed that teacher helped them find a meaningful reason to learn math during week 2}
  \item{avg_perceived_math_meaning}{Average degree to which participant agreed that teacher helped them find a meaningful reason to learn math across both weeks}
}}
\usage{
random_student_data
}
\description{
This data was randomly generated using the wakefield package on GitHub (set.seed = 1)
}
\keyword{datasets}
